chore: 32bit -> 64bit type 으로 맞추기

+++ exec/chat/chat-server/chat-server/room/Room.cpp
@@ -7,7 +7,7 @@
 
 Room::Room(
     std::shared_ptr<highp::log::Logger> logger,
-    uint32_t roomId
+    uint64_t roomId
 ) :
     _logger(std::move(logger)),
     _roomId(roomId) {
@@ -20,7 +20,7 @@ void Room::Join(const std::shared_ptr<User>& user) {
     }
 
     this->BroadcastUserJoined(
-        static_cast<uint32_t>(user->GetId()),
+        static_cast<uint64_t>(user->GetId()),
         user->GetUsername());
 
     std::scoped_lock lock{_mtx};
@@ -50,10 +50,10 @@ void Room::Leave(uint64_t userId) {
         username = (*found)->GetUsername();
         _users.erase(found);
     }
-    this->BroadcastUserLeft(static_cast<uint32_t>(userId), username);
+    this->BroadcastUserLeft(static_cast<uint64_t>(userId), username);
 }
 
-void Room::BroadcastUserJoined(uint32_t userId, std::string_view userName) {
+void Room::BroadcastUserJoined(uint64_t userId, std::string_view userName) {
     std::scoped_lock lock{_mtx};
     const auto pkt = highp::protocol::MakeUserJoinedBroadcast(userId, userName);
     for (const auto& u : _users) {
@@ -65,7 +65,7 @@ void Room::BroadcastUserJoined(uint32_t userId, std::string_view userName) {
     }
 }
 
-void Room::BroadcastUserLeft(uint32_t userId, std::string_view userName) {
+void Room::BroadcastUserLeft(uint64_t userId, std::string_view userName) {
     std::scoped_lock lock{_mtx};
     const auto pkt = highp::protocol::MakeUserLeftBroadcast(userId, userName);
     for (const auto& u : _users) {
